fix: Fix unit test failures
Relates #5065. Fix unit test failure.
